refactor excel xlsx package helpers
